work09/requirements.txt
pkg-resources==0.0.0
protobuf==3.12.2
python-memcached==1.59
six==1.15.0
